Update examples to latest code base.
    - Replace NumPy with dpnp for kernel arguments.
    - Remove the use of dpctl.device_context
