--- numba_dpex/examples/auto_offload_examples/sum-1d.py
@@ -1,48 +1,0 @@
-# Copyright 2020 - 2023 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
-
-import dpctl
-import numpy as np
-from numba import njit
-
-
-@njit
-def f1(a, b):
-    c = a + b
-    return c
-
-
-def main():
-    global_size = 64
-    local_size = 32
-    N = global_size * local_size
-    print("N", N)
-
-    a = np.ones(N, dtype=np.float32)
-    b = np.ones(N, dtype=np.float32)
-
-    print("a:", a, hex(a.ctypes.data))
-    print("b:", b, hex(b.ctypes.data))
-
-    # Use the environment variable SYCL_DEVICE_FILTER to change
-    # the default device. See
-    # https://github.com/intel/llvm/blob/sycl/sycl/doc/EnvironmentVariables.md#sycl_device_filter.
-    device = dpctl.select_default_device()
-    print("Using device ...")
-    device.print_device_info()
-
-    with dpctl.device_context(device):
-        c = f1(a, b)
-
-    print("RESULT c:", c, hex(c.ctypes.data))
-    for i in range(N):
-        if c[i] != 2.0:
-            print("First index not equal to 2.0 was", i)
-            break
-
-    print("Done...")
-
-
-if __name__ == "__main__":
-    main()
