+++ numba_dpex/examples/kernel/pairwise_distance.py
@@ -9,7 +9,7 @@ from time import time
 
 import dpctl
 import dpctl.memory as dpctl_mem
-import numpy as np
+import dpnp as np
 
 import numba_dpex as ndpx
 
@@ -54,27 +54,15 @@ def pairwise_distance(X, D, xshape0, xshape1):
 def driver():
     # measure running time
     times = list()
-
-    xbuf = dpctl_mem.MemoryUSMShared(X.size * X.dtype.itemsize)
-    x_ndarray = np.ndarray(X.shape, buffer=xbuf, dtype=X.dtype)
-    np.copyto(x_ndarray, X)
-
-    dbuf = dpctl_mem.MemoryUSMShared(D.size * D.dtype.itemsize)
-    d_ndarray = np.ndarray(D.shape, buffer=dbuf, dtype=D.dtype)
-    np.copyto(d_ndarray, D)
-
     for repeat in range(args.r):
         start = time()
         pairwise_distance[ndpx.NdRange(global_size, local_size)](
-            x_ndarray, d_ndarray, X.shape[0], X.shape[1]
+            X, D, X.shape[0], X.shape[1]
         )
         end = time()
 
         total_time = end - start
         times.append(total_time)
-
-    np.copyto(X, x_ndarray)
-    np.copyto(D, d_ndarray)
 
     return times
 
@@ -82,14 +70,7 @@ def driver():
 def main():
     times = None
 
-    # Use the environment variable SYCL_DEVICE_FILTER to change the default device.
-    # See https://github.com/intel/llvm/blob/sycl/sycl/doc/EnvironmentVariables.md#sycl_device_filter.
-    device = dpctl.select_default_device()
-    print("Using device ...")
-    device.print_device_info()
-
-    with dpctl.device_context(device):
-        times = driver()
+    times = driver()
 
     times = np.asarray(times, dtype=np.float32)
     t = Template("Average time of $runs is = ${timing}")
